+++ lib/Refactoring/CMakeLists.txt
@@ -2,6 +2,7 @@ add_swift_host_library(swiftRefactoring STATIC
   CollapseNestedIfStmt.cpp
   ConvertStringConcatenationToInterpolation.cpp
   ConvertGuardExprToIfLetExpr.cpp
+  ConvertIfLetExprToGuardExpr.cpp
   ConvertToSwitchStmt.cpp
   ConvertToTernaryExpr.cpp
   ExtractRepeatedExpr.cpp
